+++ src/main/java/qna/application/service/QnaService.java
@@ -1,12 +1,18 @@
-package qna.service;
+package qna.application.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qna.CannotDeleteException;
-import qna.NotFoundException;
-import qna.domain.*;
+import qna.domain.content.ContentType;
+import qna.domain.content.answer.Answer;
+import qna.domain.content.answer.AnswerRepository;
+import qna.domain.content.question.Question;
+import qna.domain.content.question.QuestionRepository;
+import qna.domain.log.DeleteHistory;
+import qna.domain.user.User;
+import qna.exception.CannotDeleteException;
+import qna.exception.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
