+++ src/main/java/qna/domain/content/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.content;
 
 public enum ContentType {
     QUESTION, ANSWER
