+++ src/main/java/qna/exception/UnAuthorizedException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.exception;
 
 public class UnAuthorizedException extends RuntimeException {
     private static final long serialVersionUID = 1L;
